+++ src/Tabby/Core/Base.h
@@ -22,7 +22,6 @@
 
 #ifdef TB_DEBUG
 #define TB_ENABLE_ASSERTS
-#define TB_PROFILE
 #endif
 
 #define TB_ENABLE_VERIFY